+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 * Bugfix: Don't set `muc_domain` for roomspanel if `locked_muc_domain` is `true`.
 * Bugfix: Modal auto-closes when you open it for a second time.
+* Take roster nickname into consideration when rendering messages and chat headings.
 
 ## 4.2.0 (2019-04-04)